[Refac]: 도메인 기반 폴더링 구조 정리 - project
