+++ src/main/java/org/sopt/makers/internal/controller/InternalOpenApiController.java
@@ -14,18 +14,19 @@ import org.sopt.makers.internal.dto.community.InternalCommunityPost;
 import org.sopt.makers.internal.dto.community.PostCategoryDao;
 import org.sopt.makers.internal.dto.internal.*;
 import org.sopt.makers.internal.dto.member.MemberProfileSpecificResponse;
-import org.sopt.makers.internal.dto.project.ProjectLinkDao;
+import org.sopt.makers.internal.project.domain.Project;
+import org.sopt.makers.internal.project.dto.response.ProjectLinkDao;
 import org.sopt.makers.internal.exception.ClientBadRequestException;
 import org.sopt.makers.internal.mapper.CommunityResponseMapper;
 import org.sopt.makers.internal.mapper.MemberMapper;
-import org.sopt.makers.internal.mapper.ProjectResponseMapper;
+import org.sopt.makers.internal.project.mapper.ProjectResponseMapper;
 import org.sopt.makers.internal.member.controller.coffeechat.dto.response.InternalCoffeeChatMemberResponse;
 import org.sopt.makers.internal.dto.internal.InternalMemberProjectResponse;
 import org.sopt.makers.internal.member.mapper.coffeechat.CoffeeChatResponseMapper;
 import org.sopt.makers.internal.member.repository.coffeechat.dto.InternalCoffeeChatMemberDto;
 import org.sopt.makers.internal.service.InternalApiService;
 import org.sopt.makers.internal.service.MemberService;
-import org.sopt.makers.internal.service.ProjectService;
+import org.sopt.makers.internal.project.service.ProjectService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
