+++ src/main/java/org/sopt/makers/internal/project/dto/request/ProjectUpdateRequest.java
@@ -1,15 +1,13 @@
-package org.sopt.makers.internal.dto.project;
+package org.sopt.makers.internal.project.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 
 import java.time.LocalDate;
 import java.util.List;
 
-public record ProjectSaveRequest(
+public record ProjectUpdateRequest(
         @Schema(required = true)
         String name,
-        @Schema(required = true)
-        Long writerId,
         @Schema(required = true)
         Integer generation,
         @Schema(required = true)
@@ -17,10 +15,8 @@ public record ProjectSaveRequest(
         @Schema(required = true)
         LocalDate startAt,
         LocalDate endAt,
-        @Schema(required = true)
         String[] serviceType,
         Boolean isAvailable,
-        Boolean isFounding,
         @Schema(required = true)
         String summary,
         @Schema(required = true)
@@ -30,17 +26,18 @@ public record ProjectSaveRequest(
         @Schema(required = true)
         String thumbnailImage,
         String[] images,
-        List<ProjectMemberSaveRequest> members,
-        List<ProjectLinkSaveRequest> links
+        List<ProjectMemberUpdateRequest> members,
+        List<ProjectLinkUpdateRequest> links
 ) {
-    public record ProjectMemberSaveRequest(
+    public record ProjectMemberUpdateRequest(
             Long memberId,
             String memberRole,
             String memberDescription,
             Boolean isTeamMember
     ){}
 
-    public record ProjectLinkSaveRequest(
+    public record ProjectLinkUpdateRequest(
+            Long linkId,
             String linkTitle,
             String linkUrl
     ){}
