+++ src/main/java/org/sopt/makers/internal/project/dto/response/ProjectDetailResponse.java
@@ -1,13 +1,18 @@
-package org.sopt.makers.internal.dto.project;
+package org.sopt.makers.internal.project.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 
 import java.time.LocalDate;
+import java.time.LocalDateTime;
 import java.util.List;
 
-public record ProjectUpdateRequest(
+public record ProjectDetailResponse(
+        @Schema(required = true)
+        Long id,
         @Schema(required = true)
         String name,
+        @Schema(required = true)
+        Long writerId,
         @Schema(required = true)
         Integer generation,
         @Schema(required = true)
@@ -15,8 +20,10 @@ public record ProjectUpdateRequest(
         @Schema(required = true)
         LocalDate startAt,
         LocalDate endAt,
+        @Schema(required = true)
         String[] serviceType,
         Boolean isAvailable,
+        Boolean isFounding,
         @Schema(required = true)
         String summary,
         @Schema(required = true)
@@ -26,17 +33,23 @@ public record ProjectUpdateRequest(
         @Schema(required = true)
         String thumbnailImage,
         String[] images,
-        List<ProjectMemberUpdateRequest> members,
-        List<ProjectLinkUpdateRequest> links
+        LocalDateTime createdAt,
+        LocalDateTime updatedAt,
+        List<ProjectMemberResponse> members,
+        List<ProjectLinkResponse> links
 ) {
-    public record ProjectMemberUpdateRequest(
+    public record ProjectMemberResponse(
             Long memberId,
             String memberRole,
             String memberDescription,
-            Boolean isTeamMember
+            Boolean isTeamMember,
+            String memberName,
+            List<Integer> memberGenerations,
+            String memberProfileImage,
+            Boolean memberHasProfile
     ){}
 
-    public record ProjectLinkUpdateRequest(
+    public record ProjectLinkResponse(
             Long linkId,
             String linkTitle,
             String linkUrl
