+++ src/main/java/org/sopt/makers/internal/project/dto/response/ProjectResponse.java
@@ -1,25 +1,18 @@
-package org.sopt.makers.internal.dto.project;
+package org.sopt.makers.internal.project.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 
-import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.util.List;
 
-public record ProjectDetailResponse(
+public record ProjectResponse(
         @Schema(required = true)
         Long id,
         @Schema(required = true)
         String name,
         @Schema(required = true)
-        Long writerId,
-        @Schema(required = true)
         Integer generation,
         @Schema(required = true)
         String category,
-        @Schema(required = true)
-        LocalDate startAt,
-        LocalDate endAt,
         @Schema(required = true)
         String[] serviceType,
         Boolean isAvailable,
@@ -32,23 +25,14 @@ public record ProjectDetailResponse(
         String logoImage,
         @Schema(required = true)
         String thumbnailImage,
-        String[] images,
-        LocalDateTime createdAt,
-        LocalDateTime updatedAt,
         List<ProjectMemberResponse> members,
         List<ProjectLinkResponse> links
 ) {
     public record ProjectMemberResponse(
             Long memberId,
-            String memberRole,
-            String memberDescription,
-            Boolean isTeamMember,
             String memberName,
-            List<Integer> memberGenerations,
-            String memberProfileImage,
-            Boolean memberHasProfile
+            String memberProfileImage
     ){}
-
     public record ProjectLinkResponse(
             Long linkId,
             String linkTitle,
