+++ src/main/java/org/sopt/makers/internal/project/repository/MemberProjectRelationRepository.java
@@ -1,6 +1,6 @@
-package org.sopt.makers.internal.repository;
+package org.sopt.makers.internal.project.repository;
 
-import org.sopt.makers.internal.domain.MemberProjectRelation;
+import org.sopt.makers.internal.project.domain.MemberProjectRelation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
