+++ src/main/java/org/sopt/makers/internal/service/InternalApiService.java
@@ -6,14 +6,17 @@ import org.sopt.makers.internal.domain.*;
 import org.sopt.makers.internal.dto.internal.InternalAuthVo;
 import org.sopt.makers.internal.dto.member.ActivityVo;
 import org.sopt.makers.internal.dto.member.MemberProfileProjectDao;
-import org.sopt.makers.internal.dto.project.ProjectLinkDao;
-import org.sopt.makers.internal.dto.project.ProjectMemberDao;
-import org.sopt.makers.internal.dto.project.ProjectMemberVo;
+import org.sopt.makers.internal.project.domain.Project;
+import org.sopt.makers.internal.project.dto.response.ProjectLinkDao;
+import org.sopt.makers.internal.project.dto.response.ProjectMemberDao;
+import org.sopt.makers.internal.project.dto.response.ProjectMemberVo;
 import org.sopt.makers.internal.exception.MemberHasNotProfileException;
 import org.sopt.makers.internal.exception.NotFoundDBEntityException;
 import org.sopt.makers.internal.mapper.MemberMapper;
-import org.sopt.makers.internal.mapper.ProjectMapper;
+import org.sopt.makers.internal.project.mapper.ProjectMapper;
 import org.sopt.makers.internal.member.repository.soptactivity.MemberSoptActivityRepository;
+import org.sopt.makers.internal.project.repository.ProjectQueryRepository;
+import org.sopt.makers.internal.project.repository.ProjectRepository;
 import org.sopt.makers.internal.repository.*;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
